refactor: move AdvantageColumn and Container, plus move SummaryCard as organism
